// packages/core/.tina/__generated__/types.ts
export type Maybe<T> = T | null
export type InputMaybe<T> = Maybe<T>
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K]
}
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>
}
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>
}
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  /** References another document, used as a foreign key */
  Reference: any
  JSON: any
}

export type SystemInfo = {
  __typename?: 'SystemInfo'
  filename: Scalars['String']
  title?: Maybe<Scalars['String']>
  basename: Scalars['String']
  breadcrumbs: Array<Scalars['String']>
  path: Scalars['String']
  relativePath: Scalars['String']
  extension: Scalars['String']
  template: Scalars['String']
  collection: Collection
}

export type SystemInfoBreadcrumbsArgs = {
  excludeExtension?: InputMaybe<Scalars['Boolean']>
}

export type PageInfo = {
  __typename?: 'PageInfo'
  hasPreviousPage: Scalars['Boolean']
  hasNextPage: Scalars['Boolean']
  startCursor: Scalars['String']
  endCursor: Scalars['String']
}

export type Node = {
  id: Scalars['ID']
}

export type Document = {
  id: Scalars['ID']
  _sys?: Maybe<SystemInfo>
  _values: Scalars['JSON']
}
// ...
export type Collection = {
  __typename?: 'Collection'
  name: Scalars['String']
  slug: Scalars['String']
  label?: Maybe<Scalars['String']>
  path: Scalars['String']
  format?: Maybe<Scalars['String']>
  matches?: Maybe<Scalars['String']>
  templates?: Maybe<Array<Maybe<Scalars['JSON']>>>
  fields?: Maybe<Array<Maybe<Scalars['JSON']>>>
  documents: DocumentConnection
}

export type CollectionDocumentsArgs = {
  before?: InputMaybe<Scalars['String']>
  after?: InputMaybe<Scalars['String']>
  first?: InputMaybe<Scalars['Float']>
  last?: InputMaybe<Scalars['Float']>
  sort?: InputMaybe<Scalars['String']>
}

export type DocumentNode = Config | Page | Post | Author | Tag

export type Config = Node &
  Document & {
    __typename?: 'Config'
    Placeholder?: Maybe<Scalars['String']>
    id: Scalars['ID']
    _sys: SystemInfo
    _values: Scalars['JSON']
  }

export type ConfigConnectionEdges = {
  __typename?: 'ConfigConnectionEdges'
  cursor: Scalars['String']
  node?: Maybe<Config>
}

export type ConfigConnection = Connection & {
  __typename?: 'ConfigConnection'
  pageInfo: PageInfo
  totalCount: Scalars['Float']
  edges?: Maybe<Array<Maybe<ConfigConnectionEdges>>>
}

export type PageAuthorsAuthor = Author

export type PageAuthors = {
  __typename?: 'PageAuthors'
  author?: Maybe<PageAuthorsAuthor>
}

export type PageTagsTag = Tag

export type PageTags = {
  __typename?: 'PageTags'
  tag?: Maybe<PageTagsTag>
}

export type Page = Node &
  Document & {
    __typename?: 'Page'
    date: Scalars['String']
    slug: Scalars['String']
    title: Scalars['String']
    excerpt?: Maybe<Scalars['JSON']>
    authors?: Maybe<Array<Maybe<PageAuthors>>>
    tags?: Maybe<Array<Maybe<PageTags>>>
    content: Scalars['JSON']
    id: Scalars['ID']
    _sys: SystemInfo
    _values: Scalars['JSON']
  }

export type PageConnectionEdges = {
  __typename?: 'PageConnectionEdges'
  cursor: Scalars['String']
  node?: Maybe<Page>
}

export type PageConnection = Connection & {
  __typename?: 'PageConnection'
  pageInfo: PageInfo
  totalCount: Scalars['Float']
  edges?: Maybe<Array<Maybe<PageConnectionEdges>>>
}

export type PostAuthorsAuthor = Author

export type PostAuthors = {
  __typename?: 'PostAuthors'
  author?: Maybe<PostAuthorsAuthor>
}

export type PostTagsTag = Tag

export type PostTags = {
  __typename?: 'PostTags'
  tag?: Maybe<PostTagsTag>
}

export type Post = Node &
  Document & {
    __typename?: 'Post'
    date: Scalars['String']
    slug: Scalars['String']
    title: Scalars['String']
    excerpt?: Maybe<Scalars['JSON']>
    authors?: Maybe<Array<Maybe<PostAuthors>>>
    tags?: Maybe<Array<Maybe<PostTags>>>
    content: Scalars['JSON']
    id: Scalars['ID']
    _sys: SystemInfo
    _values: Scalars['JSON']
  }

export type PostConnectionEdges = {
  __typename?: 'PostConnectionEdges'
  cursor: Scalars['String']
  node?: Maybe<Post>
}

export type PostConnection = Connection & {
  __typename?: 'PostConnection'
  pageInfo: PageInfo
  totalCount: Scalars['Float']
  edges?: Maybe<Array<Maybe<PostConnectionEdges>>>
}

export type Author = Node &
  Document & {
    __typename?: 'Author'
    name: Scalars['String']
    date: Scalars['String']
    slug: Scalars['String']
    id: Scalars['ID']
    _sys: SystemInfo
    _values: Scalars['JSON']
  }

export type AuthorConnectionEdges = {
  __typename?: 'AuthorConnectionEdges'
  cursor: Scalars['String']
  node?: Maybe<Author>
}

export type AuthorConnection = Connection & {
  __typename?: 'AuthorConnection'
  pageInfo: PageInfo
  totalCount: Scalars['Float']
  edges?: Maybe<Array<Maybe<AuthorConnectionEdges>>>
}

export type Tag = Node &
  Document & {
    __typename?: 'Tag'
    name: Scalars['String']
    date: Scalars['String']
    slug: Scalars['String']
    id: Scalars['ID']
    _sys: SystemInfo
    _values: Scalars['JSON']
  }
// ...
export type ThemeConfigNodeFragment = {
  __typename: 'Config'
  id: string
  _values: any
  _sys: {
    __typename?: 'SystemInfo'
    filename: string
    basename: string
    breadcrumbs: Array<string>
    path: string
    relativePath: string
    extension: string
  }
}

export type PostNodeFragment = {
  __typename: 'Post'
  id: string
  date: string
  slug: string
  title: string
  excerpt?: any | null
  content: any
  _sys: {
    __typename?: 'SystemInfo'
    filename: string
    basename: string
    breadcrumbs: Array<string>
    path: string
    relativePath: string
    extension: string
  }
  tags?: Array<{
    __typename: 'PostTags'
    tag?: {
      __typename: 'Tag'
      id: string
      name: string
      date: string
      slug: string
      _sys: {
        __typename?: 'SystemInfo'
        filename: string
        basename: string
        breadcrumbs: Array<string>
        path: string
        relativePath: string
        extension: string
      }
    } | null
  } | null> | null
  authors?: Array<{
    __typename: 'PostAuthors'
    author?: {
      __typename: 'Author'
      id: string
      name: string
      date: string
      slug: string
      _sys: {
        __typename?: 'SystemInfo'
        filename: string
        basename: string
        breadcrumbs: Array<string>
        path: string
        relativePath: string
        extension: string
      }
    } | null
  } | null> | null
}

export type PageNodeFragment = {
  __typename: 'Page'
  id: string
  date: string
  slug: string
  title: string
  excerpt?: any | null
  content: any
  _sys: {
    __typename?: 'SystemInfo'
    filename: string
    basename: string
    breadcrumbs: Array<string>
    path: string
    relativePath: string
    extension: string
  }
  tags?: Array<{
    __typename: 'PageTags'
    tag?: {
      __typename: 'Tag'
      id: string
      name: string
      date: string
      slug: string
      _sys: {
        __typename?: 'SystemInfo'
        filename: string
        basename: string
        breadcrumbs: Array<string>
        path: string
        relativePath: string
        extension: string
      }
    } | null
  } | null> | null
  authors?: Array<{
    __typename: 'PageAuthors'
    author?: {
      __typename: 'Author'
      id: string
      name: string
      date: string
      slug: string
      _sys: {
        __typename?: 'SystemInfo'
        filename: string
        basename: string
        breadcrumbs: Array<string>
        path: string
        relativePath: string
        extension: string
      }
    } | null
  } | null> | null
}

export type AuthorNodeFragment = {
  __typename: 'Author'
  id: string
  name: string
  date: string
  slug: string
  _sys: {
    __typename?: 'SystemInfo'
    filename: string
    basename: string
    breadcrumbs: Array<string>
    path: string
    relativePath: string
    extension: string
  }
}

export type TagNodeFragment = {
  __typename: 'Tag'
  id: string
  name: string
  date: string
  slug: string
  _sys: {
    __typename?: 'SystemInfo'
    filename: string
    basename: string
    breadcrumbs: Array<string>
    path: string
    relativePath: string
    extension: string
  }
}
// ...
export const ThemeConfigNodeFragmentDoc = gql`
  fragment ThemeConfigNode on Config {
    __typename
    _sys {
      filename
      basename
      breadcrumbs
      path
      relativePath
      extension
    }
    id
    _values
  }
`
export const PostPartsFragmentDoc = gql`
  fragment PostParts on Post {
    date
    slug
    title
    excerpt
    authors {
      __typename
      author {
        ... on Document {
          id
        }
      }
    }
    tags {
      __typename
      tag {
        ... on Document {
          id
        }
      }
    }
    content
  }
`
export const TagPartsFragmentDoc = gql`
  fragment TagParts on Tag {
    name
    date
    slug
  }
`
export const TagNodeFragmentDoc = gql`
  fragment TagNode on Tag {
    __typename
    _sys {
      filename
      basename
      breadcrumbs
      path
      relativePath
      extension
    }
    id
    ...TagParts
  }
  ${TagPartsFragmentDoc}
`
export const AuthorPartsFragmentDoc = gql`
  fragment AuthorParts on Author {
    name
    date
    slug
  }
`
export const AuthorNodeFragmentDoc = gql`
  fragment AuthorNode on Author {
    __typename
    _sys {
      filename
      basename
      breadcrumbs
      path
      relativePath
      extension
    }
    id
    ...AuthorParts
  }
  ${AuthorPartsFragmentDoc}
`
export const PostNodeFragmentDoc = gql`
  fragment PostNode on Post {
    __typename
    _sys {
      filename
      basename
      breadcrumbs
      path
      relativePath
      extension
    }
    id
    ...PostParts
    tags {
      tag {
        ... on Tag {
          ...TagNode
        }
      }
    }
    authors {
      author {
        ... on Author {
          ...AuthorNode
        }
      }
    }
  }
  ${PostPartsFragmentDoc}
  ${TagNodeFragmentDoc}
  ${AuthorNodeFragmentDoc}
`
export const PagePartsFragmentDoc = gql`
  fragment PageParts on Page {
    date
    slug
    title
    excerpt
    authors {
      __typename
      author {
        ... on Document {
          id
        }
      }
    }
    tags {
      __typename
      tag {
        ... on Document {
          id
        }
      }
    }
    content
  }
`
export const PageNodeFragmentDoc = gql`
  fragment PageNode on Page {
    __typename
    _sys {
      filename
      basename
      breadcrumbs
      path
      relativePath
      extension
    }
    id
    ...PageParts
    tags {
      tag {
        ... on Tag {
          ...TagNode
        }
      }
    }
    authors {
      author {
        ... on Author {
          ...AuthorNode
        }
      }
    }
  }
  ${PagePartsFragmentDoc}
  ${TagNodeFragmentDoc}
  ${AuthorNodeFragmentDoc}
`
export const ConfigPartsFragmentDoc = gql`
  fragment ConfigParts on Config {
    Placeholder
  }
`
export const GetConfigDocument = gql`
  query getConfig {
    config(relativePath: "index.json") {
      ...ThemeConfigNode
    }
  }
  ${ThemeConfigNodeFragmentDoc}
`
export const GetPostDocument = gql`
  query getPost($relativePath: String!) {
    post(relativePath: $relativePath) {
      ...PostNode
    }
  }
  ${PostNodeFragmentDoc}
`
export const GetPostsDocument = gql`
  query getPosts {
    postConnection {
      totalCount
      edges {
        node {
          ...PostNode
        }
      }
    }
  }
  ${PostNodeFragmentDoc}
`
export const GetPageDocument = gql`
  query getPage($relativePath: String!) {
    page(relativePath: $relativePath) {
      ...PageNode
    }
  }
  ${PageNodeFragmentDoc}
`
export const GetPagesDocument = gql`
  query getPages {
    pageConnection {
      totalCount
      edges {
        node {
          ...PageNode
        }
      }
    }
  }
  ${PageNodeFragmentDoc}
`
export const GetAuthorDocument = gql`
  query getAuthor($relativePath: String!) {
    author(relativePath: $relativePath) {
      ...AuthorNode
    }
  }
  ${AuthorNodeFragmentDoc}
`
export const GetAuthorsDocument = gql`
  query getAuthors {
    authorConnection {
      totalCount
      edges {
        node {
          ...AuthorNode
        }
      }
    }
  }
  ${AuthorNodeFragmentDoc}
`
export const GetTagDocument = gql`
  query getTag($relativePath: String!) {
    tag(relativePath: $relativePath) {
      ...TagNode
    }
  }
  ${TagNodeFragmentDoc}
`